+++ src/components/app/pageEvents/components/eventCard.tsx
@@ -1,6 +1,6 @@
 import { format } from 'date-fns'
 
-import { EventDialog } from '@/components/app/pageEvents/components/event-dialog'
+import { EventDialog } from '@/components/app/pageEvents/components/eventDialog'
 import { NextImage } from '@/components/ui/image'
 import { SWCEvents } from '@/utils/shared/getSWCEvents'
