+++ src/components/app/pageEvents/components/eventDialogContent.tsx
