+++ src/components/app/pageEvents/components/featuredEvents.tsx
