+++ src/components/app/pageEvents/components/featuredPastEvents.tsx
