+++ src/components/app/pageEvents/components/successfulEventSignupDialog.tsx
